Introduce Driver Factory
There are two versions of cAVS SST drivers: one for Debug FS interface
and one for IOCTL interface.

The IOCTL driver is used on Windows while the Debug FS is used on Linux.

This patch introduces a cAVS::DriverFactory which instantiates the correct
cAVS::Driver according to the target OS.

Change-Id: Iaa9461ad8c9dd0f080cdf31a9808f612a9a7478b
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>