Update Rest::Resource class
The Rest::Resource class is intended to be specialized by subclasses.

Each subclass has to implement a handleRequest() method which is
always the same in most of cases.

This patch updates it to introduce a default handleRequest(). The
subclasses are still free to override it if needed.

In consequences default handlers for each verb are introduced and raise
exception by default: a subclass has to override default verb handler(s)
according to those it wants to handle.

Change-Id: Ie3517fee53e524e299e50002a3632c64e15db738
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>