Rest error handler
With keep-alive feature, the tcp connection between
client and server is persistent. When the client
disconnects, an exception "ConnectionAbortedException"
is thrown at server side. This exception is caught
by the default error hander, which prints a scary log
entry.

This patch introduces an ErrorHandler that hides
this log entry, because it's a nominal case.

Change-Id: I3863129012769825520f909bde2ebbc8ef6beefd
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>