Extend build.sh test executable detection
Currently the build.sh script looks for *UnitTest* pattern in
executable files in the test output binary folder in order to
execute them to run the test sequences.

This pattern should be reduced to *Test* since we could have
unit or functional tests in these executables.

This patch updates the build.sh accordingly.

Change-Id: If2210443236d6226008cbc01b7b59a4262ef68cc
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>