Add noexcept support
Even if Visual Studio does not support C++11 noexcept keyword, we could
build the Debug Agent using other compiler such as GCC on Linux.

This patch adds a compiler definition in order to allow to use NOEXCEPT
in the project. This macro will be replaced by noexcept, except on Windows
where it will be replaced by throw().

Change-Id: Ib896b8b18614dfff98d716e8e65b8848626c4adc
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>