Conciliating windows.h and poco headers.
Including both windows.h and Poco/Foundation.h leads to conflict,
because Poco/Foundation.h already includes windows.h
in a specific way.

Therefore this patch introduces a WindowsTypes.hpp header
that should be included instead of Windows.h when a Windows
type is needed. This header complies with poco.

Change-Id: I753a0bc8a38cd3147e1a3631fd5bfa16305a9ca9
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>