Clarify http error and abort in ServerRequestHandling
The way http error and request abort are handled in ServerRequestHandling
may be confusing: an HTTP error is a response to the client when the
server cannot handle the request, while an abort is the only thing we can do when
an error occur while the server is sending the response body.

This patch applies little change to clarify it.

Change-Id: I0c55211db9ec7cfaca73cd4b23bac95c2b3ad11a
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>