Improving display of unexpected http request content
The http client simulator throws an exception when
the content returned by the server differs from the
expected one. This exception prints both obtained
content and expected content. But when the content
is very long, it's hard to find the difference.

This patch prints also the index of the first different
character, and the two substrings that differ.

Change-Id: Iffd104f1bb81d65c4f6f273d3785b1e51f76a66a
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>