Add module entries number information in logs
Currently, the log file produced by the DebugAgent is lacking
the information concerning the number of module entries, in the
cAVS Specific Header section. As a result, the module entry
table is not recognized by the FDK Tool, and is parsed as log
entries.

This patch fixes the issue by writing the number of module
entries (32 bits) before writing the module entry table.
The Unit Tests are updated according to this modification.

Change-Id: Ifaaafd7de56a0fe574286584e93bb28e4be37024
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>