Activate new URLs for the HTTP Debug Interface
Currently, the HTTP Debug Interface does not respect the
specifications described in the SwAS: there are only 3 URLs
available to interact with the underlying data model:
- /cAVS/module/entries
- /cAVS/logging/parameters
- /cAVS/logging/stream

The goal of this patch is to implement access to the data
model through the URLs described in the SwAS:
- /instance/cavs
- /instance/cavs/0
- /instance/cavs.fwlogs/0
- /type/cavs.fwlogs
- etc

This will allow the FDK Tool to correctly connect to the
DebugAgent in order to display real time logs.

This patch updates the functional tests according to the
new URLs available.

Change-Id: Ib9637fd1fa59d313eda48331f949dbd01c4bb4dc
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>