Add command line option to disable WPP interface
Currently, there is no way to enable the IOCTL control interface
of the DebugAgent without enabling the log data path through WPP
interface.

The point is that driver logs are propagated through the
same WPP session as the FW logs, and are then trashed by DebugAgent.
Since it is not possible to have several WPP real sessions at the
same time, it is not possible to get these driver logs through a
concurrent TraceView session for instance.  Driver logs are lost...

In order to work-around this limitation, this patch defines a new
command line option to enable only the interface used by
DebugAgent.

Change-Id: I870fd8f6e01174f87ad6be485f5942ef5d8e0831
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>