Remove support of old style URLs for the HTTP Debug Interface
New URLs are available to access logs parameters and logs streams.
Thus, the old URLs that are not compliant with the specifications
can be removed:
- /cAVS/logging/parameters
- /cAVS/logging/stream

The associated functional tests are removed as well by this patch.

Note: the /cAVS/module/entries is not removed because it has not
been replaced by any new URL (and this URL is useful for debug
purposes).

Change-Id: I1ef6707db3f627126c4affa001527757efbd9a67
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>