Fix functional test: debug agent shutdown while a client is consuming log
The DebugAgent functional test "debug agent shutdown while a client is
consuming log" fails sometimes in QuickBuild.
This test creates a DebugAgent instance in a thread and then creates a
HTTP client to connect to the DebugAgent, in order to send PUT and GET
commands. However, there is no synchronization between the client/server
initialization, i.e. it is possible for the client to be started before
the server ; when this happens, the client is unable to contact the
server, leading to a test failure.

This patch adds a tempo before the creation of the client, in order to
give some time to the DebugAgent to create its HTTP server (there is
currently no way to know when this server is created and ready, since
it is managed internally by the DebugAgent in a thread)

In addition, a typo between milliseconds and microseconds has been fixed.

Change-Id: Ic913bb57ca83a0b63e98058235ceb3fd38659a16
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>