Remove useless size check in tlv::TlvUnpack
The TlvUnpack::readNext() method checks the length from TLV is valid
invoking the TlvWrapperInterface::isValidSize() method.

This check is useless since the TlvWrapperInterface is supposed to raise
an exception in case of invalid length while TlvWrapperInterface::readFrom()
is invoked.

This patch removes the useless check and introduces the exception catch
instead.

Change-Id: I51629a5b12af0e7e9cda00937babec547e72de13
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>