Fixing UtilTest random crash
A test crashes randomly because the "catch" framework
macros "CHECK" are called simultaneously from several
threads, which is forbidden.

This patch uses the ability to return a result
of the std::async method, in this way it is possible
to know in the main thread if a spawned thread has failed.

Change-Id: I7c219bcc7f96830e27bf3daaa871e3a8021a119b
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>