Remove all "new" keyword to avoid memory leaks
In order to decrease risks of memory leaks, smart PTR should be used
instead of simple ptr, and the operator new shall be used only in
case there is no other solution.

This patch removes all usage of operator new when applicable. In case the
operator cannot be removed, a comment explicits why.

Change-Id: Id50f0f695cf7a18e4bf99b43ed966eaba33dd9cb
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>