Refactor Rest Response design
The current REST design allows the Resource implementation to make
error like sending two HTTP headers. The error is difficult to be
detected since it depends on code review quality.

This patch refactors the REST component in order to introduce a design
which limits the risk of such error in Resource implementations.

The main challenge is that the REST component does not know the nature
of the Resource response.

The idea is to enhance the Response class and make the Resource returning
new instance of Response. The Server is then in charge to send them,
assuring that only one HTTP header is always sent whenever an exception/
error is encountered.

The Resource are in charge of Response allocation and the Server assures
the Response is correctly sent. The usage of the unique_ptr assures that
the Response will be freed in all situation.

Few Response subclasses are introduced in order to fit Resources's
responses nature needs.

Change-Id: I9a71952cf802581b4af1ea4abc422c0db6e6e3eb
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>