Byte stream serialization robustness
Current the serializer is able to serialize structure,
but if this structure has a method toStream or fromStream,
these methods are not called.

With this patch the serializer is only able to serialize
simple integral and enum types, complex types have
to be serialized using their methods toStream or fromStream.

Note: this change has revealed some mistakes, which are
fixed in this patch.

Change-Id: I9bbee804ec652328a25fdc97efa8436de3977dc8
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>