Do not abort Debug Agent execution if PS fails
The current behavior of the Debug Agent when the Parameter Serializer
fails to start is to abort itself its execution.

This behavior is an issue since the Parameter Serializer is not needed for
all debug feature: the Debug Agent would be able to provide log
feature or topology for instance.

This patch changes the Debug Agent behavior: in case the Parameter
Serializer fails to start, the Debug Agent will continue and only debug
features which involve the Parameter Serializer won't be available.

Change-Id: I759e53839292108605dac11f0c9cebbf931406a3
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>