Fix bug regarding FW logs regression
When fetching FW log control parameters, DebugAgent triggers an IOCTL that
ends up with an error (Data buffer passed is too small).

This is due to DataSize attribute of the command header that does not
include the status size, but only the FW log input parameter body size.

This patch fixes this.

Change-Id: I5dd3a05744af878902f5bb637abae2f690cc712a
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>