New 'Locker' wrapper class: atomic method calls
This class wraps an instance of any class and ensures that each method call on
this instance is atomic w.r.t. any other call on that same instance.

This differs from ExclusiveResource, which uses the same idiom as
std::lock_guard and provides a locking sessions. As a result, it is easier to
use for the intended use-case.

Usage example:

    int f(Locker &critical)
    {
        int a = critical->method1(); /* atomic call w.r.t. the "critical" object */
        a += critical->method2(); /* atomic call w.r.t. the "critical" object */

        return a;
    }

Change-Id: I6605a8cbe666f7ffb23677a3557ca1256186a6a8
Signed-off-by: David Wagner <david.wagner@intel.com>