Require XML Schema validation on start of the PFW
Currently, the parameter-framework used by the DebugAgent doesn't
check the validity of the XML configuration files. When a user wants
to add support for a new custom module, these XML configuration
files must be updated, and some mistakes may be introduced at this
moment.
The parameter-framework is able to detect this kind of issues by
validating the XML files thanks to XSD files (XML Schemas), however
this functionality is not enabled by default.

This patch enables the XML schema validation feature of the
parameter-framework, though a new command line option --validate
or -x.
This validation will prevent the parameter-framework to start if a
mistake is found.

TrackedOn: rtc-169542
Change-Id: I1e2007bf816107a24e50a31dddf2627cb0d237aa
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>