Use strong typed parameterId
Lots of different semantic values are represented by a uint32_t
in the code. This can lead to comparing or affecting two unrelated
values.

In order to be sure that parameter ids are not
mix-up with other unrelated values, strong type it. Thus guarding against
nonsense operation at compile time.

Change-Id: I95045c24e06a9d6f3dd18f6313577c242d46e8e3
Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>