Handle moduleId greater than module_count
Currently, ModuleEntries are handled in a vector container. ModuleEntry's
module_id is given by the index of the entry in the vector.
FW does not provide the module_id information inside MODULES_INFO message.

With loadable library introduction, module_id encodes the library_id in the
most significant bits, so that it is not possible to match vector index
with module_id.

This patch modifies the lookup function to find a module entry using the
module_id as key.
Vector container implementation has been kept (map would be more
appropriate), in order to minimize the code changes, and given
the low usage of this function.

Change-Id: I95f4baf044f4a3cbf21e7a94c75e6f6382b096e4
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>