Providing a way to update easily reference data files
The reference data files (xml, html..) are used to
check functional tests. Sometimes some code changes
involve updating these reference files. But doing it
manually is painful.

This patch implements the following strategy:
Consider a reference file "module.xml". If the
associated test fails because the "got" content
differs from the reference file, a file named
"module.xml_got" will be created with the "got"
content. In this way the user can compare easily
the "got" content and the "reference" content.

If the user thinks that the "got" content should
become the reference one, he moves the file
"module.xml_got" to "module.xml" and commit it.

Change-Id: I871a44dfdd1c1ecc9c64eedea57352d806b720a4
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>