Prepare for VS 2015: Make LastError::get more reliable
Tests have shown that the last error may be reset when merely instantiating an
empty string. It means that we must call GetLastError ASAP after the potential
error. That's not bulletproof either since there could be generated code that
we don't know of but that's the best we can do.

Change-Id: I8af1afa77defb278000d3f2b3da5703af7dfda71
Signed-off-by: David Wagner <david.wagner@intel.com>