MockedDeviceCommands: templatize the ioctl description
The parameter ID determines the nature of the ioctl and must be passed as a
parameter (here, we use a template parameter) by the caller of the helper
methods. This was missed on previous patches.

Use the opportunity to add a helper for the ioctl description, which contains:
 - the size/direction (e.g. TinyGet)
 - the feature id (e.g. logs, probes)
 - the parameter id (e.g. 0, 1, ...)
 - the type of the payload

This ensures strong typing of the ioctls and help avoiding mistakes.

Change-Id: I6ffba2a2bf135a95d2aea048783238f50931e228
Signed-off-by: David Wagner <david.wagner@intel.com>