Add the parameter payload size in getModuleParameter
In order to be compliant with latest SAS, the parameter size
shall be retrieved from the parameter fw and the parameter payload
buffer shall be dimensionned correctly.
As a first step, in order not to break the windows abstraction, this
patch adds the payload size as a parameter with a default value set
to the max payload.

Change-Id: I93ccb3ce487be9ef1df0a41041dc775a43be8330
Signed-off-by: François Gaffie <francois.gaffie@intel.com>