Cpack/Windows: creating a shortcut to launch the DBGA
It is required to create a desktop shortcut that
launches the DBGA with the required command-line parameters.

Cpack/WIX allows to create desktop shortcuts, but no
way has been found to supply needed command line parameters.

Therefore this patch creates a shortcut to a batch file
that launches the DBGA with the required parameters.

Change-Id: I19bb948daead11700159aa106ff584446993529d
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>