ModuleHandlerTests: use REQUIRE instead of CHECK to avoid out-of-bounds errors
A check was trying to compare two ranges for equality but, on debug
configuration, it can cause the STL to assert if the first range is larger than
the first.

Since both sizes are checked for equality right before, we can turn that size
check to a `REQUIRE` in order to avoid spurious errors.

Change-Id: Ic81b3c71ba59aa2eb90b9057115dad90a7e7541a
Signed-off-by: David Wagner <david.wagner@intel.com>