Windows: add a manifest requiring admin rights
A manifest may either be linked into an executable or placed alongside it (in
which case in must have the same name, suffixed with '.manifest'). CMake only
gained the ability to link manifest with executables with the 3.4 version. For
now, we only require the 3.3 version which means that, for now, we have to
install the manifest with the executable.

We need this for accessing the WPP log interface.

Change-Id: Ifb3d90e2e8c55134c3c9cb73b9ea87db19437471
Signed-off-by: David Wagner <david.wagner@intel.com>