cAVS module handler refactor
The code that deserializes firmware structures is os-agnostic, but it is
currently performed in the os-specific part.

This patch moves this code in the os-agnostic part, using a way that minimizes
API impacts (tests are not modified). But this way is not the ideal one, because
the layers of the cAVS component are a little blurry. A deeper subsequent
refactor is still needed.

Change-Id: I1803827ae2bb5ae1b97b7655a26a4fdd1ebc1135
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>