Util::BlockingQueue: ability to open the queue
BlockingQueue was opened by default, then it can
be closed but not opened again.

This patch provides an open() method that allows
to open/close the queue multiple times. When
the queue is opened, content is cleared. That's
why the clear() method is moved to the private part,
it shall not be called directly.

Change-Id: Ibad3936a4955560724ccb3ed58344c19d9aafc05
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>