Probing: deriving System handles
Event handles can be injected for test purpose.
But it is not possible to subclass them in order to
add test-specific behavior required by probe
functional tests.

This patch allows to subclass event handles.

This patch renames also the EventHandle::get() method
to EventHandle::handle() in order to not mixup it with
std::unique_ptr::get().

Change-Id: I355a3d2dd6742def03e758b246ab1efe4461e3f8
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>