HttpClientSimulator: treating reference files as binary files
Previously test reference files were treated as text files,
and were trimmed before comparison.

But this behaviour doesn't work for binary files, and probing
requires binary files.

So this patch treats reference files as binary files, because
it works for both text and binary files. This involves to update
reference files, because they are not trimmed anymore.

Change-Id: I158c81fe8fbd7d32b8b82f516be0d8649fb2ccdd
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>