DBGA: Linux adaptation: refactor SystemDevice
This patch refactors SystemDevice to support multi client.
Only one command can be run at a time. Client will wait to acquire
the mutex as a first implementation.
It also provides a higher level function to read/write a command
in an atomic way. It simplifies the interactions with the system device.

Change-Id: Iffb251d45510e0225a2c4b91e21feff4efd96250
Signed-off-by: François Gaffie <francois.gaffie@intel.com>