cAVS/Prober: tinyset ioctl with dynamic params
Currently the tinyset ioctl assumes that feature
param id is static, so it is supplied through
a template parameter.

Unfortunately for probe injection, the feature
probe id becomes dynamic (its value depends of
the probe index).

Therefore this patch provides a new tinyset api
with dynamic parameters, which is used by the
old template-based api.

Change-Id: I1a85d96bff39111dfe13f9ee00458c15978d4181
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>