Add a missing static member definition
A static member was declared in-class with a constant-initializer. That does
not constitute a definition and that member may not be used as a reference.
In order to do that, the member must be defined in a namespace scope without an
initializer.

It was working fine in Release configuration on Linux and on both Debug and
Release configurations on Windows, probably thanks to inlining but it was
broken on Linux in Debug configuration.

See https://gcc.gnu.org/wiki/VerboseDiagnostics#missing_static_const_definition

Change-Id: I93db332e36a6ddb6f80ba735210d32ff3d8b3fd1
Signed-off-by: David Wagner <david.wagner@intel.com>