Moving ProbePointId structure in fw part and using it in packets
Currently there is 2 versions of the ProbePointId structure into
the DBGA:
- one in the driver/windows part
- one is the os-agnosic probe part

But actually this structure is defined by the
FW, not the driver. So this patch defines only one
ProbePointId structure in the DBGA/Fw part, which is
used by the driver part. Note: Data validation is
now performed by ProbePointId setters.

This patch also uses this structure into the probe
extraction packets, because the "probePointId" packet
member has been misunderstood: it is not the probe index,
but the associated probe point id (module type, module
instance, pin type, pin index).

Finally this patch changes the "Extractor" class
to dispatch packets to probe streams according to
the probe point id (instead of the probe index).

Change-Id: I19e60934c8fe899c59c4b62ec37df0337f9101a8
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>