Add a missing static member definition
A static member was declared in-class with a constant-initializer. That does
not constitute a definition and that member may not be used as a reference.
In order to do that, the member must be defined in a namespace scope.

It was working fine in Release configuration on Linux and on both Debug and
Release configurations on Windows, probably thanks to inlining but it was
broken on Linux in Debug configuration.

Change-Id: I8e221b902bbd8230cf111f9c55e92db22c084aef
Signed-off-by: François Gaffie <francois.gaffie@intel.com>