Introducing util::RingBuffer
A ring buffer is needed to store data comming from
http layer until it is injected to the driver.

A first possibility consists in using the existing
RingBufferReader/RingBufferWriter classes, but
these classes are dedicated to work with a buffer
that is read/written by a third party (the driver),
involving that overflow/underflow can happen.

In our case overflow/underflow will never happen:
- if the ring buffer is full, data retrieving from
  the http layer must be stopped
- if the ring buffer is empty, silence must
  be injected to the driver
That's why RingBufferReader/Writer classes are not
suitable for this need.

This pach introduces a ringbuffer class that guarantees
that overflow/underflow cannot happen.

Write/Read methods exist in blocking/non blocking
versions.

Change-Id: Ia7dbcadb92f8271824b6025c864072dbf3fb460c
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>