Introducing Injector
The injector is an active object that forwards data
from an input ring buffer (feed by the http
layer) to an output ring buffer (consumed by the driver).

If the input ring buffer is empty, silence is written
to the output ring buffer. To achieve it, the
injector knows the audio format sample byte size in
order to not break the sample alignment.

Change-Id: I6df923ede2d1ab8a0128caa16f95a8d1735da2a8
Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>