Perf/Linux: implement get/setState calls to the hardware
The DBGA directly modifies the firmware property holding the perf measurement
state but it also needs to increment the power management ref counter to prevent
the DSP from sleeping. Caveat: only the counter of core 0 is incremented, as it
is supposed to be sufficient for keep the whole DSP awaken.

Change-Id: I9f18b37a8ea26746b2259d1a9c7f15488fc722c2
Signed-off-by: David Wagner <david.wagner@intel.com>