Adding missing info performance data in KCPS http response.
According to cAVS FDK 1.0 Feature Debug SwAS 1.00 section 5.16.2.1,
the XML answer of HTTP request for performance data should contain
parameter is_removed, currently it is not the case.

This patch add this parameter to the XML answer to conform to the SwAS.
The functionnal test have to be modified in the same time in order
to pass compilation & test.

Change-Id: I29fdc8e0d5489ea086135969b35b3cf224da033c
Signed-off-by: Louis LE GALL <louis.le.gall@intel.com>