Linux adaptation: enforce log test case
Logger unit test is not thread safe right now. A stop can be called
before a wait if the logger is started / stopped immediately.
This patch enforces the test by udating the mock to take into
account that the stop may be called before wait.
Wait mock will behave as the non mocked compress device and return
an IoException if stop is called before.

Change-Id: Ibddb54341a6f453946b1e0aaba70e76d2832b652
Signed-off-by: François Gaffie <francois.gaffie@intel.com>