Move all remaining module-handling from System to ModuleHandler
This is the next step in moving responsibilities out of the System class and
reducing its interface.

The ModuleHandler now owns the hw/fw configs ; as a result, its methods no
longer need to be given the number of config items are to be retrieved since
that information is known by the ModuleHandler (e.g. the number of gateways).
It greatly simplifies its API.

The tests are heavily reworked because of that: instead of injecting these
numbers in an ad-hoc function, we must build a complete sample environment.  In
the same time, the test cases can be factorized in a more BDD-oriented fashion.
Next, factorizing linux and windows tests shall be rather straightforward.

The hw and fw configs as well as the module type entries are now cached by the
ModuleHandler (they used to be cached by the System class). They are not
supposed to change dynamically at run-time.

Change-Id: I87423541396ef2a58c254bd2907821e7dd74dbba
Signed-off-by: David Wagner <david.wagner@intel.com>